clients/python-client/requirements.txt
grpcio==1.73.0
grpcio-tools==1.73.0
protobuf==6.31.1
